refactored `KeyExecutives` for `stocks.fa.mgmt`
