--- openbb_sdk/providers/fmp/openbb_fmp/models/executives.py
@@ -1,98 +1,0 @@
-"""FMP Key Executives Fetcher."""
-
-
-from datetime import datetime
-from typing import Any, Dict, List, Optional
-
-from openbb_provider.abstract.data import Data
-from openbb_provider.abstract.fetcher import Fetcher
-from openbb_provider.abstract.query_params import QueryParams
-from openbb_provider.models.base import BaseSymbol
-from openbb_provider.models.key_executives import (
-    KeyExecutivesData,
-    KeyExecutivesQueryParams,
-)
-
-from openbb_fmp.utils.helpers import get_data
-
-
-class FMPKeyExecutivesQueryParams(QueryParams, BaseSymbol):
-    """FMP Key Executives QueryParams.
-
-    Source: https://financialmodelingprep.com/developer/docs/#Key-Executives
-
-    Parameter
-    ---------
-    symbol : str
-        The symbol of the company.
-    key_executive_name : Optional[str]
-        The name of the key executive.
-    key_executive_title : Optional[str]
-        The title of the key executive.
-    key_executive_title_since : Optional[datetime]
-        The title since of the key executive.
-    key_executive_year_born : Optional[datetime]
-        The year born of the key executive.
-    key_executive_gender: Optional[str]
-        The gender of the key executive.
-    """
-
-    key_executive_name: Optional[str]
-    key_executive_title: Optional[str]
-    key_executive_title_since: Optional[datetime]
-    key_executive_year_born: Optional[datetime]
-    key_executive_gender: Optional[str]
-
-
-class FMPKeyExecutivesData(Data):
-    name: str
-    title: str
-    titleSince: Optional[datetime]
-    yearBorn: Optional[datetime]
-    gender: Optional[str]
-
-
-class FMPKeyExecutivesFetcher(
-    Fetcher[
-        KeyExecutivesQueryParams,
-        KeyExecutivesData,
-        FMPKeyExecutivesQueryParams,
-        FMPKeyExecutivesData,
-    ]
-):
-    @staticmethod
-    def transform_query(params: Dict[str, Any]) -> FMPKeyExecutivesQueryParams:
-        return FMPKeyExecutivesQueryParams(**params)
-
-    @staticmethod
-    def extract_data(
-        query: FMPKeyExecutivesQueryParams, credentials: Optional[Dict[str, str]]
-    ) -> List[FMPKeyExecutivesData]:
-        api_key = credentials.get("fmp_api_key") if credentials else ""
-
-        base_url = "https://financialmodelingprep.com/api/v3/"
-        request_url = f"{base_url}key-executives/{query.symbol}?apikey={api_key}"
-        data = get_data(request_url)
-        if isinstance(data, dict):
-            raise ValueError("Expected list of dicts, got dict")
-
-        if query.key_executive_name:
-            data = [x for x in data if x["name"] == query.key_executive_name]
-        if query.key_executive_title:
-            data = [x for x in data if x["title"] == query.key_executive_title]
-        if query.key_executive_title_since:
-            data = [
-                x for x in data if x["titleSince"] == query.key_executive_title_since
-            ]
-        if query.key_executive_year_born:
-            data = [x for x in data if x["yearBorn"] == query.key_executive_year_born]
-        if query.key_executive_gender:
-            data = [x for x in data if x["gender"] == query.key_executive_gender]
-
-        return [FMPKeyExecutivesData(**d) for d in data]
-
-    @staticmethod
-    def transform_data(
-        data: List[FMPKeyExecutivesData],
-    ) -> List[FMPKeyExecutivesData]:
-        return data
