+++ openbb_sdk/sdk/provider/openbb_provider/models/key_executives.py
@@ -1,59 +1,26 @@
 """Key Executives Data Model."""
 
 
-from datetime import datetime
 from typing import Optional
 
 from openbb_provider.abstract.data import Data
 from openbb_provider.abstract.query_params import QueryParams
 from openbb_provider.models.base import BaseSymbol
 
+from pydantic import Field
+
 
 class KeyExecutivesQueryParams(QueryParams, BaseSymbol):
-    """Key executives query.
-
-    Parameter
-    ---------
-    symbol : str
-        The symbol of the company.
-    key_executive_name : Optional[str]
-        The name of the key executive to be retrieved.
-    key_executive_title : Optional[str]
-        The title of the key executive to be retrieved.
-    key_executive_title_since : Optional[datetime]
-        The period since the position of the key executive to be retrieved.
-    key_executive_year_born : Optional[datetime]
-        The year born of the key executive to be retrieved.
-    key_executive_gender: Optional[str]
-        The gender of the key executive to be retrieved.
-    """
-
-    key_executive_name: Optional[str]
-    key_executive_title: Optional[str]
-    key_executive_title_since: Optional[datetime]
-    key_executive_year_born: Optional[datetime]
-    key_executive_gender: Optional[str]
+    """Key Executives Query."""
 
 
 class KeyExecutivesData(Data):
-    """Key Executives Data.
+    """Key Executives Data."""
 
-    Returns
-    -------
-    name : Optional[str]
-        The name of the key executive.
-    title : Optional[str]
-        The title of the key executive.
-    title_since : Optional[datetime]
-        The title since of the key executive.
-    year_born : Optional[datetime]
-        The year born of the key executive.
-    gender : Optional[str]
-        The gender of the key executive.
-    """
-
-    name: Optional[str]
-    title: Optional[str]
-    title_since: Optional[datetime]
-    year_born: Optional[datetime]
-    gender: Optional[str]
+    title: str = Field(description="Designation of the key executive.")
+    name: str = Field(description="Name of the key executive.")
+    pay: Optional[int] = Field(description="Pay of the key executive.")
+    currency_pay: str = Field(description="The currency of the pay.")
+    gender: Optional[str] = Field(description="Gender of the key executive.")
+    year_born: Optional[str] = Field(description="Birth year of the key executive.")
+    title_since: Optional[int] = Field(description="Date the tile was held since.")
